fix: move function into scripts
